requirements.txt
dill
h5py==2.8.0
imutils
Keras==2.3.1
Markdown
matplotlib
numpy
pandas==1.2.4
Pillow
scikit-image==0.14.2
scikit-learn==0.22.1
scipy
seaborn
six
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tqdm
opencv-python